+++ src/main/java/com/dwarfeng/dct/util/FlatDataUtil.java
@@ -2,6 +2,7 @@ package com.dwarfeng.dct.util;
 
 import com.dwarfeng.dct.bean.dto.FlatData;
 import com.dwarfeng.dutil.basic.time.TimeUtil;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 
 import java.time.Instant;
 import java.util.Objects;
@@ -46,6 +47,26 @@ public final class FlatDataUtil {
 
         flatData.setHappenedDate(TimeUtil.toDate(instant));
         flatData.setHappenedDateNanoOffset(TimeUtil.toNanoOffset(instant));
+    }
+
+    /**
+     * 根据指定参数构造新的扁平数据实例。
+     *
+     * <p>
+     * 该方法会根据指定的瞬时时间同步设置发生时间与毫秒内纳秒偏移。
+     *
+     * @param pointKey        指定的点位主键。
+     * @param value           指定的值。
+     * @param happenedInstant 指定的发生瞬时时间。
+     * @return 新构造的扁平数据实例。
+     * @throws NullPointerException 入口参数为 <code>null</code>。
+     * @since 2.0.1
+     */
+    public static FlatData newInstance(LongIdKey pointKey, String value, Instant happenedInstant) {
+        Objects.requireNonNull(pointKey, "pointKey 不能为空");
+        Objects.requireNonNull(happenedInstant, "happenedInstant 不能为空");
+
+        return new FlatData(pointKey, value, TimeUtil.toDate(happenedInstant), TimeUtil.toNanoOffset(happenedInstant));
     }
 
     private FlatDataUtil() {